Renamed two files and added the correct raw counts file
I renamed two files we are using so it is clear that at step 1 we use the FPKM data and not the raw counts that we have to use at step 2. Moreover, now, the correct raw counts file that Brian uploaded is in the code at step 2. Lastly, I removed the lines about integers and negative values since there aren't any in the correct raw counts file.