Code update with DESeq for histology and gender completely and partly for performance and age
Created new files where we can put the DEGs for every specific variable. For smoking, gender and histology they are in there. I am not sure for age what to choose as reference so still need to figure that out and then continue with the analysis. For performance there are a lot of groups so maybe first make the decision which groups you want to compare before continuing.